requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
pillow==11.2.1
psycopg2==2.9.10
sqlparse==0.5.3
uWSGI==2.0.29
